Move funcion for pulse-skipping to module
